Update README and refactor CLI commands for improved functionality
- Updated Node.js version requirement in README to v18+.
- Added import functionality to CLI for importing multiple project folders.
- Refactored command imports to streamline command handling.
- Introduced new remove command with confirmation prompt for deleting projects.
- Enhanced help command to reflect updated commands and usage.
- Created core utility functions for color logging and configuration management.
- Established a new ConfigStore class for managing project entries.
