+++ src/commands/index.ts
@@ -1,0 +1,8 @@
+export { runConfig } from "./config";
+export { runVersion } from "./version";
+export { runOpen } from "./open";
+export { runList } from "./list";
+export { runHelp } from "./help";
+export { runRemove } from "./remove";
+export { runUpdate } from "./update";
+export { runImport } from "./import";
