+++ src/commands/version.ts
@@ -3,9 +3,9 @@ import fs from "node:fs";
 import readline from "node:readline";
 import { spawn } from "node:child_process";
 
-import { version } from "package" assert { type: "json" };
+import { ConfigStore } from "~/core";
 
-import type { ConfigStore } from "~/configStore";
+import { version } from "package" assert { type: "json" };
 
 function saveVersionCheckData(store: ConfigStore, data: Record<string, any>) {
   try {
@@ -24,7 +24,7 @@ function loadVersionCheckData(store: ConfigStore) {
     );
     return parsed && typeof parsed === "object" ? parsed : {};
   } catch (err: any) {
-    console.warn("Failed to read version check file. Resetting.", err.message);
+    console.warn("Failed to read version check file. removeting.", err.message);
     return {};
   }
 }
